refactor: resolve review feedback, flatten transport layer, and revert JsonRpc visibility
